+++ ui/CHANGES.md
@@ -1,7 +1,7 @@
 #### Version: 1.4.0 (LibGDX 1.9.6)
 - **Added**: [#259](https://github.com/kotcrab/vis-editor/issues/259) `VisImageTextButton#setGenerateDisabledImage(boolean)` along with getter, added `VisImageButton#isGenerateDisabledImage`
 - **Fixed**: [#255](https://github.com/kotcrab/vis-editor/issues/255) Custom cursor handling in `VisSplitPane` and `MultiSplitPane` 
-- `Spinner` now implements `Disableable`
+- [#252](https://github.com/kotcrab/vis-editor/issues/252) `Spinner` now implements `Disableable`
 
 #### Version: 1.3.0 (LibGDX 1.9.6)
 - **Added**: `VisUI#dispose (boolean disposeSkin)`